+++ lib/auto_form_faker/railtie.rb
@@ -4,6 +4,13 @@ module AutoFormFaker
       ActiveSupport.on_load(:action_view) do
         include AutoFormFaker::FormHelperExtension
       end
+      
+      # Include SimpleForm extension if SimpleForm is available
+      config.after_initialize do
+        if defined?(SimpleForm)
+          SimpleForm::FormBuilder.include(AutoFormFaker::SimpleFormExtension)
+        end
+      end
     end
   end
 end